+++ CDTDatastore/touchdb/ChangeTracker/TDURLConnectionChangeTracker.m
@@ -3,7 +3,7 @@
 //  
 //
 //  Created by Adam Cox on 1/5/15.
-//  Copyright (c) 2015 IBM.
+//  Copyright © 2015, 2017 IBM Corp. All rights reserved.
 //
 //  Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file
 //  except in compliance with the License. You may obtain a copy of the License at
@@ -40,6 +40,9 @@
 @property (nonatomic, strong) CDTURLSession * session;
 @property (nonatomic, strong) CDTURLSessionTask * task;
 @end
+
+static const int kChangeQueueThreshold = 500;
+static const float kChangeQueuePollingRate = 0.1f;
 
 @implementation TDURLConnectionChangeTracker
 
@@ -281,8 +284,23 @@ didReceiveChallenge:(NSURLAuthenticationChallenge *)challenge
     }
     
     [self clearConnection];
-    
-    if (restart){
+
+    if (restart) {
+        // Throttle the rate at which we get the list of changes. If we already have more than
+        // kChangeQueueThreshold changes to be processed, wait until we fall below that threshold
+        // before we get any more. Note that this is not a hard limit, and the number of changes may
+        // exceed kChangeQueueThreshold, but it won't go vastly above the threshold. This saves us
+        // from consuming large amounts of memory by allocating a TDPulledRevision for each
+        // change we are waiting to pull and keeps our peak memory usage much smaller during
+        // pulls of large numbers of changes.
+        if ([_client respondsToSelector:@selector(sizeOfChangeQueue)]) {
+            while ([_client sizeOfChangeQueue] > kChangeQueueThreshold &&
+                   [[NSRunLoop currentRunLoop]
+                          runMode:NSDefaultRunLoopMode
+                       beforeDate:[NSDate dateWithTimeIntervalSinceNow:kChangeQueuePollingRate]])
+                ;
+        }
+
         [self start];  // Next poll...
     } else {
         [self stopped];