+++ CDTDatastore/touchdb/TDReplicator.m
@@ -6,7 +6,7 @@
 //  Copyright (c) 2011 Couchbase, Inc. All rights reserved.
 //
 //  Modifications for this distribution by Cloudant, Inc., Copyright (c) 2014 Cloudant, Inc.
-//  Copyright © 2016 IBM Corporation. All rights reserved.
+//  Copyright © 2016, 2017 IBM Corporation. All rights reserved.
 //
 //  Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file
 //  except in compliance with the License. You may obtain a copy of the License at
@@ -309,6 +309,11 @@ NSString* TDReplicatorStartedNotification = @"TDReplicatorStarted";
             dispatch_group_leave(taskGroup);
         }
     }
+
+    [self.session finishTasksAndInvalidate];
+    self.session = nil;
+    self.sessionConfigDelegate = nil;
+    self.interceptors = nil;
 }