+++ CDTDatastore/touchdb/TD_Database.h
@@ -6,7 +6,15 @@
  *  Copyright (c) 2011 Couchbase, Inc. All rights reserved.
  *
  *  Modifications for this distribution by Cloudant, Inc., Copyright (c) 2014 Cloudant, Inc.
+ *  Copyright © 2017 IBM Corporation. All rights reserved.
  *
+ *  Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file
+ *  except in compliance with the License. You may obtain a copy of the License at
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *  Unless required by applicable law or agreed to in writing, software distributed under the
+ *  License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND,
+ *  either express or implied. See the License for the specific language governing permissions
+ *  and limitations under the License.
  */
 
 #import "TD_Revision.h"
@@ -234,6 +242,8 @@ extern const TDChangesOptions kDefaultTDChangesOptions;
     format ("designdocname/viewname"), it attempts to load the view properties from the
     design document and compile them with the TDViewCompiler. */
 - (TD_View*)compileViewNamed:(NSString*)name status:(TDStatus*)outStatus;
+
+- (void)clearPendingAttachments;
 
 @property (readonly) NSArray* allViews;